Factor out shared test functions.
